backend/brick_detector.py
import cv2
import numpy as np

class BrickDetector:
    def __init__(self):
        #TODO: Initialize ML model here
        #self.model = load_your_model()
        pass
    
    def detect_bricks(self, image_path):
        """
        Detect and classify Lego bricks in an image
        This is a simplified example - replace with your actual CV/ML pipeline
        """
        #Reads image
        image = cv2.imread(image_path)
        if image is None:
            raise ValueError("Could not read image")
        
        #Converts to RGB
        image_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        
        # TODO: Implement actual detection logic
        # Example steps:
        # 1. Preprocess image (resize, normalize)
        # 2. Run through object detection model
        # 3. Classify detected objects
        # 4. Count instances of each brick type
        
        #Mock implementation
        results = self._mock_detection(image_rgb)
        
        return results
    
    def _mock_detection(self, image):
        """Mock detection for development"""
        #Need to implement:
        # - YOLO/SSD for object detection
        # - CNN for brick classification
        # - Color detection algorithms
        
        height, width = image.shape[:2]
        
        #Simulates some detections
        mock_results = [
            {
                "id": "3001",
                "name": "2x4 Brick",
                "color": self._detect_color(image, 100, 100),
                "quantity": 1,
                "confidence": 0.95,
                "bbox": [100, 100, 200, 150]
            },
           
        ]
        
        return mock_results
    
    def _detect_color(self, image, x, y):
        """Simple color detection (mock implementation)"""
        colors = ["Red", "Blue", "Yellow", "Green", "Black", "White"]
        return np.random.choice(colors)

#Usage example
detector = BrickDetector()